Feat: Add method for cloning a repo, close #5
