+++ test/test_CI_integration_tests.py
@@ -2,7 +2,6 @@
 import pytest
 
 import CI
-from CI.CI_helpers import *
 
 
 @pytest.fixture
@@ -21,17 +20,3 @@ def test_index(client):
     res = client.get('/')
     assert res.status_code == 200
     assert b'Hello, World!' == res.data
-
-
-def test_run_commands_positive():
-    commands = ["echo hello", "echo bye"]
-    command_output = run_commands(commands)
-    assert command_output == [b'hello\n', b'bye\n']
-
-
-def test_run_commands_negative():
-    """ First command is invalid, will throw an exception and 'Error' is appended
-    to list """
-    commands = ["ech hello", "echo hello"]
-    command_output = run_commands(commands)
-    assert command_output == ["Error", b'hello\n']
